Added a continuous python executor
This can be reworked to run anything, not necessarily python code.
i.e. You can change it so that it executes a bash command instead
of executing a python file.

It listens to a file, and whenever it is modified it executes it as
a python script.

But that could be changed to execute a specified command whenever it
is modified.