Make ssh-agent check only on non-tmux terminal
Checking if the ssh-agent is running and that it has the desired keys
is expensive and it only needs to be done once.

Since ssh-agent will stay up for the rest of the session with our keys,
we can do it only once on the initial terminal before attaching to the
tmux session, but every subsequent pane won't check that again.