at-office.el/llm-goose-gen: Fix failure message due to bad interactive block.
The `(interactive)` block was covering the entire function body, which of course resulted in an error since `interactive` blocks are supposed to fill in the parameters via a list, in this case the input-prompt parameter.

The fix was to make `(interactive ...)` essentially return 1 item in the list that represents the user prompt resulting from a `(read-string ...)` call.